2024.09.18 update: StringBuilder replaced with String concat
